+++ server/json/user/list.go
@@ -1,6 +1,6 @@
 package user
 
-import "github.com/nepp-tumsat/documents-api/model"
+import "github.com/nepp-tumsat/documents-api/infrastructure/model"
 
 type UserListResponse struct {
 	Users []UserGetResponse `json:"users"`
@@ -10,7 +10,7 @@ func ToUserListResponse(users []model.User) UserListResponse {
 	var jsonUsers []UserGetResponse
 	var jsonUser UserGetResponse
 	for _, user := range users {
-		jsonUser = ToUserGetResponse(user)
+		jsonUser = UserGetResponse{UserName: user.UserName}
 		jsonUsers = append(jsonUsers, jsonUser)
 	}
 	return UserListResponse{Users: jsonUsers}
